Split node.go file up further
