+++ cluster/raft/node.go
@@ -1,9 +1,6 @@
 package raft
 
 import (
-	"bytes"
-	"context"
-	"encoding/binary"
 	"fmt"
 	"log"
 	"net/http"
@@ -92,30 +89,6 @@ type node struct {
 // Name implements [process.Runnable.Name].
 func (n *node) Name() string {
 	return fmt.Sprintf("raft node %d", n.conf.ID)
-}
-
-func (n *node) AsPeer() cluster.Peer {
-	return cluster.Peer{
-		ID:   n.conf.ID,
-		Addr: n.addr,
-	}
-}
-
-func (n *node) Handler() http.Handler {
-	h := new(Handler)
-
-	h.node = n
-
-	mux := http.NewServeMux()
-	mux.Handle("/message", http.HandlerFunc(h.messageHandler))
-
-	h.Handler = mux
-
-	return h
-}
-
-func (n *node) Receive(msg raftpb.Message) error {
-	return n.raft.Step(context.TODO(), msg)
 }
 
 // Name implements [process.Runnable.Run].
@@ -247,65 +220,4 @@ func (n *node) Status() raft.Status {
 
 func (n *node) Tick() {
 	n.manualTick <- time.Now()
-}
-
-func (n *node) Handle(t cluster.ProposalType, f cluster.ProposalFunc) {
-	if f := n.proposalHandlers[t]; f != nil {
-		panic(fmt.Errorf("%w: %d", cluster.ErrDuplicateProposalType, t))
-	}
-	n.proposalHandlers[t] = f
-}
-
-func encodeProposal(id uint64, t uint32, data []byte) []byte {
-	header := make([]byte, 12)
-	binary.LittleEndian.PutUint64(header, id)
-	binary.LittleEndian.PutUint32(header[8:], uint32(t))
-
-	buf := new(bytes.Buffer)
-	buf.Write(header)
-	buf.Write(data)
-	return buf.Bytes()
-}
-
-func (n *node) Propose(t cluster.ProposalType, data []byte) (resp any, err error) {
-	id, await := n.proposalReporter.Await()
-	defer n.proposalReporter.Close(id)
-
-	enc := encodeProposal(id, uint32(t), data)
-	if err := n.raft.Propose(context.TODO(), enc); err != nil {
-		return nil, err
-	}
-
-	result := <-await
-	return result.resp, result.err
-}
-
-func decodeProposal(enc []byte) (id uint64, t uint32, data []byte) {
-	id = binary.LittleEndian.Uint64(enc[0:8])
-	t = binary.LittleEndian.Uint32(enc[8:12])
-	data = enc[12:]
-	return
-}
-
-func (n *node) applyProposal(data []byte) {
-	id, pt, dec := decodeProposal(data)
-
-	f, ok := n.proposalHandlers[cluster.ProposalType(pt)]
-	if !ok {
-		panic(fmt.Errorf("%w: %d", cluster.ErrUnknownProposalType, pt))
-	}
-
-	resp, err := f(dec)
-
-	send, ok := n.proposalReporter.Send(id)
-	if ok {
-		send <- result{resp, err}
-	} else if err != nil {
-		log.Println("application error applying proposal:", err)
-	}
-}
-
-type result struct {
-	resp any
-	err  error
 }
