+++ cluster/raft/node_handler.go
@@ -1,12 +1,26 @@
 package raft
 
 import (
+	"context"
 	"io"
 	"net/http"
 
 	"github.com/golang/protobuf/proto"
 	"go.etcd.io/raft/v3/raftpb"
 )
+
+func (n *node) Handler() http.Handler {
+	h := new(Handler)
+
+	h.node = n
+
+	mux := http.NewServeMux()
+	mux.Handle("/message", http.HandlerFunc(h.messageHandler))
+
+	h.Handler = mux
+
+	return h
+}
 
 type Handler struct {
 	http.Handler
@@ -42,4 +56,8 @@ func (h *Handler) postMessageHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.WriteHeader(http.StatusOK)
+}
+
+func (n *node) Receive(msg raftpb.Message) error {
+	return n.raft.Step(context.TODO(), msg)
 }
